governance/paperclip/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-ready-paperclip-workflows"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pyyaml>=6.0.1"
]

[project.scripts]
paperclip-daily-standup = "paperclip_workflows.daily_standup:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
